+++ src/main/java/duke/Duke.java
@@ -1,6 +1,11 @@
 package duke;
 
-import command.Command;
+import duke.command.Command;
+import duke.exception.DukeException;
+import duke.parser.Parser;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
 
 public class Duke {
     private Storage storage;
