+++ src/main/java/duke/command/Command.java
@@ -1,8 +1,8 @@
-package command;
+package duke.command;
 
-import duke.TaskList;
-import duke.Ui;
-import duke.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+import duke.storage.Storage;
 
 public abstract class Command {
     public abstract void execute(TaskList tasks, Ui ui, Storage storage);
