+++ src/main/java/duke/command/ListCommand.java
@@ -1,8 +1,8 @@
-package command;
+package duke.command;
 
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
 
 public class ListCommand extends Command {
     @Override
