+++ src/main/java/duke/command/UnmarkCommand.java
@@ -1,11 +1,15 @@
-package command;
+package duke.command;
 
-import duke.*;
+import duke.exception.DukeException;
+import duke.storage.Storage;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.ui.Ui;
 
-public class MarkCommand extends Command {
+public class UnmarkCommand extends Command {
     private final String details;
 
-    public MarkCommand(String details) {
+    public UnmarkCommand(String details) {
         if (details.isBlank()) {
             throw new DukeException("☹ OOPS!!! The description of a new task cannot be empty.");
         }
@@ -17,9 +21,9 @@ public class MarkCommand extends Command {
         try {
             int i = Integer.parseInt(this.details);
             Task taskToMark = tasks.get(i-1);
-            taskToMark.markUnmark(true);
+            taskToMark.markUnmark(false);
             storage.update(tasks);
-            ui.show("Nice! I've marked this task as done:");
+            ui.show("OK, I've marked this task as not done yet:");
             ui.show(String.valueOf(taskToMark));
         } catch (NumberFormatException err) {
             throw new DukeException("☹ OOPS!!! " + this.details + " is not a valid integer for indexing the task list.");
