+++ src/main/java/duke/storage/Storage.java
@@ -1,4 +1,7 @@
-package duke;
+package duke.storage;
+
+import duke.exception.DukeException;
+import duke.task.*;
 
 import java.io.*;
 import java.time.LocalDateTime;
